Fix warnings when calling StructureListBox drawItem

+++ OPHD/UI/StructureListBox.cpp
@@ -158,9 +158,9 @@ void StructureListBox::update()
 	{
 		drawItem(renderer, *static_cast<StructureListBoxItem*>(mItems[i]),
 			positionX(),
-			positionY() + (i * LIST_ITEM_HEIGHT),
-			static_cast<float>(item_width()),
-			static_cast<float>(draw_offset()),
+			positionY() + (static_cast<int>(i) * LIST_ITEM_HEIGHT),
+			item_width(),
+			draw_offset(),
 			i == currentSelection());
 	}
